may update message na

+++ App/controllers/ListingController.php
@@ -4,7 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
-use VARIANT;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -20,10 +21,7 @@ class ListingController
     public function index()
     {
 
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
-
-
-
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', ['listings' => $listings]);
     }
@@ -76,7 +74,7 @@ class ListingController
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -124,6 +122,9 @@ class ListingController
 
             $this->db->query($query, $newListingData);
 
+            Session::setFlashMessage('success_message', 'Listing created successfully');
+
+
             redirect('/listings');
         }
     }
@@ -143,14 +144,23 @@ class ListingController
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        //Check if listing exist
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
         }
+
+        //Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+            return redirect('/listings/' . $listing->id);
+        }
+
+
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
         //Set flash message
-        $_SESSION['success_message'] = 'Listing delete successfully';
+        Session::setFlashMessage('success_message', 'Listing delete successfully');
 
         redirect('/listings');
     }
@@ -237,15 +247,15 @@ class ListingController
             foreach (array_keys($updatedValues) as $field) {
                 $updateFields[] = "{$field} = :{$field}";
             }
+
             $updateFields = implode(', ', $updateFields);
 
             $updateQuery = "UPDATE listings SET $updateFields WHERE id = :id";
 
             $updatedValues['id'] = $id;
-
             $this->db->query($updateQuery, $updatedValues);
 
-            $_SESSION['success_message'] = 'Listing Updated';
+            Session::setFlashMessage('success_message', 'Listing update successfully');
 
             redirect('/listings/' . $id);
         }
